requirements.txt
pillow==10.4.0
opencv-python==4.10.0.84
transformers==4.42.3
torch==2.3.1
python-dotenv==1.0.1
pandas==2.2.2
streamlit==1.36.0
